Use i64 instead of u64 for group ring IDs
This is easier to store in an SQLite table, which uses i64 for all
integer columns. It's also more consistent across platforms, since
Java also has no u64 equivalent.

+++ src/rust/src/android/android_platform.rs
@@ -998,7 +998,7 @@ impl Platform for AndroidPlatform {
                 return;
             }
         };
-        let ring_id = u64::from(ring_id) as jlong;
+        let ring_id = jlong::from(ring_id);
         let sender = match env.byte_array_from_slice(&sender) {
             Ok(slice) => JObject::from(slice),
             Err(error) => {
